processor/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
deprecation==2.1.0
filelock==3.17.0
flask==3.1.1
frozenlist==1.6.0
fsspec==2025.3.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mlx==0.23.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
postgrest==1.0.2
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-core==2.33.2
pyjwt==2.10.1
pytest==8.3.5
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
realtime==2.4.3
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
strenum==0.4.15
supabase==2.15.2
supafunc==0.9.4
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.3.0
websockets==14.2
werkzeug==3.1.3
yarl==1.20.0
